+++ full-moon/src/ast/mod.rs
@@ -421,6 +421,11 @@ pub enum Stmt<'a> {
     #[display(fmt = "{}", _0)]
     While(While<'a>),
 
+    /// A compound assignment, such as `+=`
+    /// Only available when the "roblox" feature flag is enabled
+    #[cfg(feature = "roblox")]
+    #[display(fmt = "{}", _0)]
+    CompoundAssignment(CompoundAssignment<'a>),
     /// A continue statement
     /// Only available when the "roblox" feature flag is enabled.
     #[cfg(feature = "roblox")]
@@ -1918,23 +1923,6 @@ impl<'a> FunctionDeclaration<'a> {
     pub fn with_body(self, body: FunctionBody<'a>) -> Self {
         Self { body, ..self }
     }
-}
-
-macro_rules! make_op {
-    ($enum:ident, $(#[$outer:meta])* { $($operator:ident,)+ }) => {
-        #[derive(Clone, Debug, Display, PartialEq, Owned, Node, Visit)]
-        #[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
-        #[visit(skip_visit_self)]
-        $(#[$outer])*
-        #[display(fmt = "{}")]
-        pub enum $enum<'a> {
-            #[cfg_attr(feature = "serde", serde(borrow))]
-            $(
-                #[allow(missing_docs)]
-                $operator(Cow<'a, TokenReference<'a>>),
-            )+
-        }
-    };
 }
 
 make_op!(BinOp,